Inject details about asset and normalized quantities in published events
